backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
python-jose[cryptography]
bcrypt==4.0.1
python-multipart
groq
pypdf2
reportlab
requests